refactor: moved parsing helper methods
